requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
PyJWT==2.6.0
requests==2.28.1
lxml~=4.9.1
PyMySQL==1.0.3
DBUtils~=3.0.3